# openapi.json
{
  "components": {
    "schemas": {
      "DatabaseInfo": {
        "additionalProperties": false,
        "properties": {
          "database": {
            "examples": [
              "app_db"
            ],
            "type": "string"
          },
          "destinations": {
            "examples": [
              [
                "local_storage",
                "offsite_s3"
              ]
            ],
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "engine": {
            "examples": [
              "postgres"
            ],
            "type": "string"
          },
          "host": {
            "examples": [
              "pg-cluster.internal"
            ],
            "type": "string"
          },
          "is_running": {
            "examples": [
              false
            ],
            "type": "boolean"
          },
          "last_run": {
            "$ref": "#/components/schemas/Run"
          },
          "name": {
            "examples": [
              "production_pg"
            ],
            "type": "string"
          },
          "next_run_at": {
            "format": "date-time",
            "type": "string"
          },
          "port": {
            "examples": [
              5432
            ],
            "format": "int64",
            "type": "integer"
          },
          "schedule": {
            "examples": [
              "0 */6 * * *"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "engine",
          "host",
          "port",
          "database",
          "schedule",
          "destinations",
          "is_running"
        ],
        "type": "object"
      },
      "DestinationInfo": {
        "additionalProperties": false,
        "properties": {
          "bucket": {
            "examples": [
              "company-backups"
            ],
            "type": "string"
          },
          "encrypted": {
            "examples": [
              true
            ],
            "type": "boolean"
          },
          "endpoint": {
            "examples": [
              "s3.amazonaws.com"
            ],
            "type": "string"
          },
          "name": {
            "examples": [
              "offsite_s3"
            ],
            "type": "string"
          },
          "path": {
            "examples": [
              "backups/{{ .Database }}.enc"
            ],
            "type": "string"
          },
          "retention": {
            "$ref": "#/components/schemas/RetentionPolicy"
          },
          "type": {
            "examples": [
              "s3"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "type",
          "path",
          "encrypted"
        ],
        "type": "object"
      },
...
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ErrorModel.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
      "HistoryOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/HistoryOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/Run"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "runs",
          "total"
        ],
        "type": "object"
      },
      "ListBackupsOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ListBackupsOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "backups": {
            "items": {
              "$ref": "#/components/schemas/StorageBackupItem"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "backups"
        ],
        "type": "object"
      },
      "RestoreBackupInputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/RestoreBackupInputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "backup_path": {
            "description": "Path of backup file in destination",
            "type": "string"
          },
          "destination": {
            "description": "Name of storage destination",
            "type": "string"
          }
        },
        "required": [
          "destination",
          "backup_path"
        ],
        "type": "object"
      },
      "RestoreBackupOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/RestoreBackupOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "duration_ms": {
            "examples": [
              1250
            ],
            "format": "int64",
            "type": "integer"
          },
          "logs": {
            "type": "string"
          },
          "message": {
            "examples": [
              "Database dev_db restored successfully"
            ],
            "type": "string"
          }
        },
        "required": [
          "message",
          "duration_ms"
        ],
        "type": "object"
      },
...
      "Run": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/Run.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "completed_at": {
            "format": "date-time",
            "type": "string"
          },
          "database_name": {
            "type": "string"
          },
          "destinations": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "duration_ms": {
            "format": "int64",
            "type": "integer"
          },
          "engine": {
            "type": "string"
          },
          "error_message": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "logs": {
            "type": "string"
          },
          "size_bytes": {
            "format": "int64",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "database_name",
          "engine",
          "status",
          "started_at",
          "duration_ms",
          "size_bytes",
          "destinations"
        ],
        "type": "object"
      },
      "RunLogsOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/RunLogsOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "id": {
            "examples": [
              "run-123"
            ],
            "type": "string"
          },
          "logs": {
            "examples": [
              "[2026-09-14T00:00:00Z] Backup started..."
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "logs"
        ],
        "type": "object"
      },
...
      "TriggerBackupOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/TriggerBackupOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "message": {
            "examples": [
              "Backup triggered"
            ],
            "type": "string"
          },
          "run_id": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "title": "arkbase API",
    "version": "1.0.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/databases": {
      "get": {
        "operationId": "list-databases",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatabaseInfo"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List configured databases",
        "tags": [
          "Databases"
        ]
      }
    },
    "/api/v1/databases/{id}/backup": {
      "post": {
        "operationId": "trigger-database-backup",
        "parameters": [
          {
            "description": "Name of the database to backup",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Name of the database to backup",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerBackupOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Trigger on-demand backup for a database",
        "tags": [
          "Databases"
        ]
      }
    },
    "/api/v1/databases/{id}/backups": {
      "get": {
        "operationId": "list-database-backups",
        "parameters": [
          {
            "description": "Name of the database",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Name of the database",
              "type": "string"
            }
          },
          {
            "description": "Optional filter by storage destination name",
            "explode": false,
            "in": "query",
            "name": "destination",
            "schema": {
              "description": "Optional filter by storage destination name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListBackupsOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List backup files in storage destinations for a database",
        "tags": [
          "Databases"
        ]
      }
    },
    "/api/v1/databases/{id}/restore": {
      "post": {
        "operationId": "restore-database-backup",
        "parameters": [
          {
            "description": "Name of the database to restore",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Name of the database to restore",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreBackupInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreBackupOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Restore a database from a storage destination backup",
        "tags": [
          "Databases"
        ]
      }
    },
    "/api/v1/destinations": {
      "get": {
        "operationId": "list-destinations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DestinationInfo"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List configured backup destinations",
        "tags": [
          "Destinations"
        ]
      }
    },
    "/api/v1/history": {
      "get": {
        "operationId": "get-history",
        "parameters": [
          {
            "description": "Number of records to return",
            "explode": false,
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 20,
              "description": "Number of records to return",
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "description": "Offset for pagination",
            "explode": false,
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "description": "Offset for pagination",
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "description": "Filter by database name",
            "explode": false,
            "in": "query",
            "name": "database",
            "schema": {
              "description": "Filter by database name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Get backup run history",
        "tags": [
          "History"
        ]
      }
    },
    "/api/v1/history/{id}": {
      "get": {
        "operationId": "get-run-detail",
        "parameters": [
          {
            "description": "Run UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Run UUID",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Get run execution details and logs",
        "tags": [
          "History"
        ]
      }
    },
    "/api/v1/history/{id}/logs": {
      "get": {
        "operationId": "get-run-logs",
        "parameters": [
          {
            "description": "Run UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Run UUID",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunLogsOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Get run execution output logs",
        "tags": [
          "History"
        ]
      }
    },
    "/api/v1/stats": {
      "get": {
        "operationId": "get-stats",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryStats"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Get overall backup statistics",
        "tags": [
          "System"
        ]
      }
    },
    "/api/v1/status": {
      "get": {
        "operationId": "get-status",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Get daemon status",
        "tags": [
          "System"
        ]
      }
    }
  }
}